Add 'deleteWithoutEditor' property to enable deleting cell contents without opening/closing the editor
Actually, instead of refactoring the cell's value validation and persistence code from CellEditor.js to a shared module and reference it from CellEditing.js, we just let the cell editor open normally and, if it's value is ok, we automatically close it without user interaction. The editor stays open if the validation fails.